#14 remove the LLM-API-Module and move the LLM into the actual application.
