--- team_a/LLM-API-Modules/lib/llm_api_modules.dart
@@ -1,8 +1,0 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
-library;
-
-export 'src/llm_api_modules_base.dart';
-
-// TODO: Export any libraries intended for clients of this package.
