--- team_a/LLM-API-Modules/pubspec.yaml
@@ -1,16 +1,0 @@
-name: llm_api_modules
-description: A starting point for Dart libraries or applications.
-version: 1.0.0
-# repository: https://github.com/my_org/my_repo
-
-environment:
-  sdk: ^3.5.3
-
-# Add regular dependencies here.
-dependencies:
-  # path: ^1.8.0
-   http: ^1.2.1 
-
-dev_dependencies:
-  lints: ^4.0.0
-  test: ^1.24.0
