+++ team_a/teamA/lib/Api/llm_api.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 import 'dart:convert';
 import 'package:http/http.dart' as http;
+import 'package:learninglens_app/services/api_service.dart';
 
 class LlmApi 
 {
@@ -43,7 +44,7 @@ class LlmApi
       Uri url, Map<String, String> postHeaders, Object postBody) async 
   {
     final httpPackageResponse =
-        await http.post(url, headers: postHeaders, body: postBody);
+        await ApiService().httpPost(url, headers: postHeaders, body: postBody);
 
     if (httpPackageResponse.statusCode != 200) {
       print('Failed to retrieve the http package!');
