+++ team_a/teamA/lib/Views/edit_questions.dart
@@ -1,11 +1,10 @@
-import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:learninglens_app/Controller/custom_appbar.dart';
 import 'package:learninglens_app/Views/send_quiz_to_moodle.dart';
+import 'package:learninglens_app/llm/openai_api.dart';
 import 'package:learninglens_app/services/local_storage_service.dart';
 import '../Api/moodle_api_singleton.dart';
 import 'package:learninglens_app/Controller/beans.dart';
 import 'package:flutter/material.dart';
-import 'package:llm_api_modules/openai_api.dart';
 import 'quiz_generator.dart';
 
 class EditQuestions extends StatefulWidget {
