+++ team_a/teamA/lib/Views/essay_generation.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:learninglens_app/Api/moodle_api_singleton.dart';
 import 'package:learninglens_app/Controller/custom_appbar.dart';
 import 'package:learninglens_app/Views/essay_edit_page.dart';
+import 'package:learninglens_app/llm/claudeai_api.dart';
 import 'package:learninglens_app/services/local_storage_service.dart';
 import 'dart:convert';
 import '../Api/llm_api.dart';
-import 'package:llm_api_modules/openai_api.dart';
-import 'package:llm_api_modules/claudeai_api.dart';
+import 'package:learninglens_app/llm/openai_api.dart';
+
 
 // Required Components:
 // 2 Dropdowns: 1 for the Grade Level and 1 for the Point Scale
