+++ team_a/teamA/lib/Views/view_submissions.dart
@@ -1,14 +1,13 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:learninglens_app/Controller/custom_appbar.dart';
+import 'package:learninglens_app/llm/claudeai_api.dart';
+import 'package:learninglens_app/llm/openai_api.dart';
 import 'package:learninglens_app/services/local_storage_service.dart';
 import '../Api/moodle_api_singleton.dart';
 import 'package:learninglens_app/Controller/beans.dart';
 import 'view_submission_detail.dart';
 import '../Api/llm_api.dart';
 import 'dart:convert';
-import 'package:llm_api_modules/openai_api.dart';
-import 'package:llm_api_modules/claudeai_api.dart';
 import 'package:intl/intl.dart';
 
 class SubmissionList extends StatefulWidget {
