+++ team_a/teamA/lib/llm/openai_api.dart
@@ -1,41 +1,55 @@
 import 'dart:async';
 import 'dart:convert';
-import 'package:http/http.dart' as http;
+import 'package:learninglens_app/services/api_service.dart';
 
-class ClaudeAiAPI {
-  final String claudeAiKey;
-  ClaudeAiAPI(this.claudeAiKey);
+class OpenAiLLM {
+  final String openAiKey;
+  OpenAiLLM(this.openAiKey);
 
   Map<String, dynamic> convertHttpRespToJson(String httpResponseString) {
     return (json.decode(httpResponseString) as Map<String, dynamic>);
   }
 
+  ///
+  ///
+  ///
   String getPostBody(String queryMessage) {
     return jsonEncode({
-      'model': 'claude-3-5-sonnet-20240620',
-      'max_tokens': 1024,
-      'system': 'Be precise and concise',
+      'model': 'gpt-4o-mini',
       'messages': [
+        {'role': 'system', 'content': 'Be precise and concise'},
         {'role': 'user', 'content': queryMessage}
       ]
     });
   }
 
+  ///
+  ///
+  ///
   Map<String, String> getPostHeaders() {
     return ({
-      'anthropic-version': '2023-06-01',
+      'accept': 'application/json',
       'content-type': 'application/json',
-      'anthropic-dangerous-direct-browser-access': 'true',
-      'x-api-key': claudeAiKey,
+      'authorization': 'Bearer $openAiKey',
     });
   }
 
-  Uri getPostUrl() => Uri.https('api.anthropic.com', '/v1/messages');
+  ///
+  ///
+  ///
+  Uri getPostUrl() => Uri.https('api.openai.com', '/v1/chat/completions');
 
+  ///
+  ///
+  ///
   Future<String> postMessage(
       Uri url, Map<String, String> postHeaders, Object postBody) async {
     final httpPackageResponse =
-        await http.post(url, headers: postHeaders, body: postBody);
+        await ApiService().httpPost(url, headers: postHeaders, body: postBody);
+
+    print(url);
+    print(postHeaders);
+    print(postBody);
 
     if (httpPackageResponse.statusCode != 200) {
       print('Failed to retrieve the http package!');
@@ -45,6 +59,7 @@ class ClaudeAiAPI {
       return "";
     }
 
+    print("In postmessage : ${httpPackageResponse.body}");
     return httpPackageResponse.body;
   }
 
@@ -74,6 +89,9 @@ class ClaudeAiAPI {
     return parsedResp;
   }
 
+  ///
+  ///
+  ///
   Future<String> postToLlm(String queryPrompt) async {
     var resp = "";
 
@@ -85,6 +103,9 @@ class ClaudeAiAPI {
     return resp;
   }
 
+  ///
+  ///
+  ///
   Future<String> queryAI(String query) async {
     final postHeaders = getPostHeaders();
     final postBody = getPostBody(query);
@@ -97,10 +118,10 @@ class ClaudeAiAPI {
         convertHttpRespToJson(httpPackageRespString);
 
     var retResponse = "";
-    for (var respChoice in httpPackageResponseJson['content']) {
-      retResponse += respChoice['text'];
+    for (var respChoice in httpPackageResponseJson['choices']) {
+      retResponse += respChoice['message']['content'];
     }
-    // print("In queryAI - content :  $retResponse");
+    print("In queryAI - content :  $retResponse");
     return retResponse;
   }
 }
