+++ team_a/teamA/pubspec.yaml
@@ -9,8 +9,6 @@ environment:
   sdk: ^3.1.1
 
 dependencies:
-  llm_api_modules:
-    path: ../LLM-API-Modules
   flutter:
     sdk: flutter
